fix: FK type mismatch on payment_memo_invoices + improve setup route output
- Change payment_memo_id from unsignedBigInteger to unsignedInteger to match payment_memos.id (increments = unsigned INT)
- Add early output, Throwable catch, and explicit Content-Type header to setup-production route

+++ database/migrations/2026_05_08_100002_create_payment_memo_invoices_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
 
         Schema::create('payment_memo_invoices', function (Blueprint $table) {
             $table->increments('id');
-            $table->unsignedBigInteger('payment_memo_id');
+            $table->unsignedInteger('payment_memo_id');
             $table->unsignedInteger('invoice_id');
             $table->decimal('grand_total', 15, 2);     // snapshot saat memo dibuat
             $table->decimal('sisa_tagihan', 15, 2);    // snapshot sisa bayar saat memo dibuat
